Added casts from uint8_t to int

+++ IOT/lib/controll/decoder.c
@@ -74,7 +74,7 @@ void decoder_decode(const char *message) {
         {
             temp_hum[i] = temp_hum_ptr[i];
         }
-        int temperature = temp_hum[0] * 10 + temp_hum[1]; //Constructing temperature from response array
+        int temperature = (int)temp_hum[0] * 10 + (int)temp_hum[1]; //Constructing temperature from response array
         decoder_send(message, RES_GID_SEN_VAL, 4, temperature);
     }
     else if (t0_is_req == 0 && t2_is_get == 0 && t3_is_hum == 0)
@@ -85,7 +85,7 @@ void decoder_decode(const char *message) {
         {
             temp_hum[i] = temp_hum_ptr[i];
         }
-        int humidity = temp_hum[2] * 10 + temp_hum[3]; //Constructing humidity from response array
+        int humidity = (int)temp_hum[2] * 10 + (int)temp_hum[3]; //Constructing humidity from response array
         decoder_send(message, RES_GID_SEN_VAL, 4, humidity);
     }
     //A request for light sensor reading
